create & use ExceptionPage for 404s, remove LinkButton
